pyproject.toml
[project]
name = "create-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[anthropic]>=1.2.0",
    "langchain-openai>=1.1.6",
    "pydantic>=2.12.5",
    "ijson>=3.4.0.post0",
    "sqlalchemy>=2.0.45",
    "fastapi>=0.128.0",
]
